changed sbyte so it can handle bigger boards

+++ Virus/UnitTesting/Board.cs
@@ -10,15 +10,15 @@ namespace UnitTesting
 {
     public class Board
     {
-        public sbyte[,] board;
+        public int[,] board;
         public int boardSize { get; set; }
 
-        public sbyte playerTurn;
+        public int playerTurn;
         public bool jumping = false;
         private VirusPlayer player1, player2;
         public Board(int size)
         {
-            board = new sbyte[size, size];
+            board = new int[size, size];
             boardSize = size;
         }
         public void StartGame()
@@ -33,9 +33,9 @@ namespace UnitTesting
         /// Returns the number of points for all players starting with player 1 (0 indexed)
         /// </summary>
         /// <returns></returns>
-        internal sbyte[] GetScore()
+        internal int[] GetScore()
         {
-            sbyte[] result = new sbyte[2];
+            int[] result = new int[2];
             for (int x = 0; x < boardSize; x++)
             {
                 for (int y = 0; y < boardSize; y++)
@@ -55,7 +55,7 @@ namespace UnitTesting
 
         internal void reset()
         {
-            board = new sbyte[boardSize, boardSize];
+            board = new int[boardSize, boardSize];
             board[0, 0] = 1;
             board[boardSize - 1, boardSize - 1] = 1;
             board[boardSize - 1, 0] = 2;
@@ -97,9 +97,9 @@ namespace UnitTesting
         /// <param name="brickToMove"></param>
         /// <param name="moveToHere"></param>
         /// <returns></returns>
-        public virtual sbyte MoveBrick(sbyte brickToMoveX, sbyte brickToMoveY, sbyte moveToHereX, sbyte moveToHereY)
+        public virtual int MoveBrick(int brickToMoveX, int brickToMoveY, int moveToHereX, int moveToHereY)
         {
-            sbyte move = Move(playerTurn, brickToMoveX, brickToMoveY, moveToHereX, moveToHereY);
+            int move = Move(playerTurn, brickToMoveX, brickToMoveY, moveToHereX, moveToHereY);
 
             if (move != -1)
             {
@@ -130,7 +130,7 @@ namespace UnitTesting
 
         private bool CantMove()
         {
-            List<Tuple<sbyte, sbyte, sbyte, sbyte>> moves = new List<Tuple<sbyte, sbyte, sbyte, sbyte>>();
+            List<Tuple<int, int, int, int>> moves = new List<Tuple<int, int, int, int>>();
             foreach (var item in GetBricks(playerTurn))
             {
 
@@ -138,23 +138,23 @@ namespace UnitTesting
                 {
                     for (int y = -1; y <= 1; y++)
                     {
-                        sbyte x2 = -1;
-                        sbyte y2 = -1;
+                        int x2 = -1;
+                        int y2 = -1;
                         if (x > item.Item2)
-                            x2 = (sbyte)(x - item.Item2);
+                            x2 = (int)(x - item.Item2);
                         else
-                            x2 = (sbyte)(item.Item2 - x);
+                            x2 = (int)(item.Item2 - x);
                         if (y > item.Item3)
-                            y2 = (sbyte)(y - item.Item3);
+                            y2 = (int)(y - item.Item3);
                         else
-                            y2 = (sbyte)(item.Item3 - y);
+                            y2 = (int)(item.Item3 - y);
 
-                        sbyte result = TryMakeMove(playerTurn, item.Item2, item.Item3, x2, y2);
+                        int result = TryMakeMove(playerTurn, item.Item2, item.Item3, x2, y2);
                         if (result != -1)
                         {
-                            if (!moves.Contains(new Tuple<sbyte, sbyte, sbyte, sbyte>(item.Item2, item.Item3, x2, y2)))
+                            if (!moves.Contains(new Tuple<int, int, int, int>(item.Item2, item.Item3, x2, y2)))
                             {
-                                moves.Add(new Tuple<sbyte, sbyte, sbyte, sbyte>(item.Item2, item.Item3, x2, y2));
+                                moves.Add(new Tuple<int, int, int, int>(item.Item2, item.Item3, x2, y2));
                             }
                         }
                     }
@@ -171,16 +171,16 @@ namespace UnitTesting
         /// The method returns a list with the player number followed by the index of the brick
         /// </summary>
         /// <returns></returns>
-        public List<Tuple<sbyte, sbyte, sbyte>> GetBricks()
+        public List<Tuple<int, int, int>> GetBricks()
         {
-            List<Tuple<sbyte, sbyte, sbyte>> result = new List<Tuple<sbyte, sbyte, sbyte>>();
+            List<Tuple<int, int, int>> result = new List<Tuple<int, int, int>>();
             for (int x = 0; x < boardSize; x++)
             {
                 for (int y = 0; y < boardSize; y++)
                 {
                     if (board[x, y] != 0)
                     {
-                        result.Add(new Tuple<sbyte, sbyte, sbyte>(board[x, y], (sbyte)x, (sbyte)y));
+                        result.Add(new Tuple<int, int, int>(board[x, y], (int)x, (int)y));
                     }
                 }
             }
@@ -191,22 +191,22 @@ namespace UnitTesting
         /// to the playerNumber
         /// </summary>
         /// <returns></returns>
-        public List<Tuple<sbyte, sbyte, sbyte>> GetBricks(int playerNumber)
+        public List<Tuple<int, int, int>> GetBricks(int playerNumber)
         {
-            List<Tuple<sbyte, sbyte, sbyte>> result = new List<Tuple<sbyte, sbyte, sbyte>>();
+            List<Tuple<int, int, int>> result = new List<Tuple<int, int, int>>();
             for (int x = 0; x < boardSize; x++)
             {
                 for (int y = 0; y < boardSize; y++)
                 {
                     if (board[x, y] != 0 && board[x, y] == playerNumber)
                     {
-                        result.Add(new Tuple<sbyte, sbyte, sbyte>((sbyte)playerNumber, (sbyte)x, (sbyte)y));
+                        result.Add(new Tuple<int, int, int>((int)playerNumber, (int)x, (int)y));
                     }
                 }
             }
             return result;
         }
-        protected sbyte Move(sbyte playerNumber, sbyte brickToMoveX, sbyte brickToMoveY, sbyte moveToHereX, sbyte moveToHereY)
+        protected int Move(int playerNumber, int brickToMoveX, int brickToMoveY, int moveToHereX, int moveToHereY)
         {
             if (IsOnTheBoard(moveToHereX, moveToHereY))
                 if (IsAPiece(playerNumber, brickToMoveX, brickToMoveY))
@@ -216,7 +216,7 @@ namespace UnitTesting
             return -1;
         }
 
-        private bool IsAPiece(sbyte playerNumber, sbyte brickToMoveX, sbyte brickToMoveY)
+        private bool IsAPiece(int playerNumber, int brickToMoveX, int brickToMoveY)
         {
             if (board[brickToMoveX, brickToMoveY] == playerNumber)
                 return true;
@@ -231,7 +231,7 @@ namespace UnitTesting
         /// <param name="moveToHereX"></param>
         /// <param name="moveToHereY"></param>
         /// <returns></returns>
-        public sbyte TryMakeMove(sbyte playerNumber, sbyte brickToMoveX, sbyte brickToMoveY, sbyte moveToHereX, sbyte moveToHereY)
+        public int TryMakeMove(int playerNumber, int brickToMoveX, int brickToMoveY, int moveToHereX, int moveToHereY)
         {
             if (!IsAPiece(playerNumber, brickToMoveX, brickToMoveY))
                 return -1;
@@ -242,8 +242,8 @@ namespace UnitTesting
             if (!IsOnTheBoard(moveToHereX, moveToHereY))
                 return -1;
 
-            sbyte player = board[brickToMoveX, brickToMoveY];
-            sbyte taken = 0;
+            int player = board[brickToMoveX, brickToMoveY];
+            int taken = 0;
 
             //Jump move
             if (jumping)
@@ -262,7 +262,7 @@ namespace UnitTesting
             return taken;
         }
 
-        private bool IsOnTheBoard(sbyte moveToHereX, sbyte moveToHereY)
+        private bool IsOnTheBoard(int moveToHereX, int moveToHereY)
         {
             try
             {
@@ -278,11 +278,11 @@ namespace UnitTesting
             return true;
         }
 
-        private sbyte MakeMove(sbyte brickToMoveX, sbyte brickToMoveY, sbyte moveToHereX, sbyte moveToHereY)
+        private int MakeMove(int brickToMoveX, int brickToMoveY, int moveToHereX, int moveToHereY)
         {
-            sbyte player = board[brickToMoveX, brickToMoveY];
+            int player = board[brickToMoveX, brickToMoveY];
 
-            sbyte taken = 0;
+            int taken = 0;
 
             //Jump move
             if (jumping)
@@ -300,9 +300,9 @@ namespace UnitTesting
             board[moveToHereX, moveToHereY] = player;
             return taken;
         }
-        private sbyte AssumeCapturedPieces(sbyte moveToHereX, sbyte moveToHereY, sbyte player)
+        private int AssumeCapturedPieces(int moveToHereX, int moveToHereY, int player)
         {
-            sbyte taken = 0;
+            int taken = 0;
             try
             {
                 if (board[moveToHereX - 1, moveToHereY - 1] != 0 && board[moveToHereX - 1, moveToHereY - 1] != player)
@@ -396,9 +396,9 @@ namespace UnitTesting
             }
             return taken;
         }
-        private sbyte CapturePieces(sbyte moveToHereX, sbyte moveToHereY, sbyte player)
+        private int CapturePieces(int moveToHereX, int moveToHereY, int player)
         {
-            sbyte taken = 0;
+            int taken = 0;
             try
             {
                 if (board[moveToHereX - 1, moveToHereY - 1] != 0 && board[moveToHereX - 1, moveToHereY - 1] != player)
@@ -501,7 +501,7 @@ namespace UnitTesting
             return taken;
         }
 
-        private bool IsMoveEligable(sbyte brickToMoveX, sbyte brickToMoveY, sbyte moveToHereX, sbyte moveToHereY)
+        private bool IsMoveEligable(int brickToMoveX, int brickToMoveY, int moveToHereX, int moveToHereY)
         {
             if (IsJumpEligable(brickToMoveX, brickToMoveY, moveToHereX, moveToHereY))
             {
@@ -516,7 +516,7 @@ namespace UnitTesting
             return false;
         }
 
-        private bool IsNormalMoveEligable(sbyte brickToMoveX, sbyte brickToMoveY, sbyte moveToHereX, sbyte moveToHereY)
+        private bool IsNormalMoveEligable(int brickToMoveX, int brickToMoveY, int moveToHereX, int moveToHereY)
         {
             for (int i = -1; i <= 1; i++)
             {
@@ -624,7 +624,7 @@ namespace UnitTesting
             return false;
         }
 
-        private bool IsJumpEligable(sbyte brickToMoveX, sbyte brickToMoveY, sbyte moveToHereX, sbyte moveToHereY)
+        private bool IsJumpEligable(int brickToMoveX, int brickToMoveY, int moveToHereX, int moveToHereY)
         {
             if (brickToMoveX - moveToHereX == 0 && brickToMoveY - moveToHereY == 2 || brickToMoveX - moveToHereX == 0 && brickToMoveY - moveToHereY == -2)
                 return true;
@@ -637,7 +637,7 @@ namespace UnitTesting
             return false;
         }
 
-        private bool IsNotOccupied(sbyte moveToHereX, sbyte moveToHereY)
+        private bool IsNotOccupied(int moveToHereX, int moveToHereY)
         {
             try
             {
